marketing/requirements.txt
# Database connectivity
psycopg2-binary==2.9.9

# Data manipulation and analysis
pandas==2.2.0
numpy==1.26.3

# Plotting and visualization
matplotlib==3.8.2
seaborn==0.13.1

# Configuration
configparser==6.0.0
